dom/base/test/test_bug738108.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=738108
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 738108</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=738108">Mozilla Bug 738108</a>
<p id="display"></p>
<div id="content" style="display: none">
  <div id="foo"></div>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 738108 **/
is(document.querySelector("#foo"), $("foo"),
   "querySelector on document should find element by id")
is($("content").querySelector("#foo"), $("foo"),
   "querySelector on parent element should find element by id")
is($("foo").querySelector("#foo"), null,
   "querySelector on element should not find the element itself by id")

is(document.querySelectorAll("#foo").length, 1,
   "querySelectorAll on document should find element by id")
is($("content").querySelectorAll("#foo").length, 1,
   "querySelectorAll on parent element should find element by id")
is($("foo").querySelectorAll("#foo").length, 0,
   "querySelectorall on element should not find the element itself by id")

</script>
</pre>
</body>
</html>
